some test cleanup
