+++ kompendium-core/src/test/kotlin/org/leafygreens/kompendium/util/TestModels.kt
@@ -42,18 +42,12 @@ data class TestResponse(val c: String)
 
 data class TestCreatedResponse(val id: Int, val c: String)
 
-object TestDeleteResponse
-
 data class ComplexRequest(
   val org: String,
   @KompendiumField("amazing_field")
   val amazingField: String,
   val tables: List<NestedComplexItem>
-) {
-  fun testThing() {
-    println("hey mom 👋")
-  }
-}
+)
 
 data class NestedComplexItem(
   val name: String,
@@ -75,10 +69,9 @@ data class DefaultParameter(
   @KompendiumParam(ParamType.PATH) val c: Boolean
 )
 
-sealed class TestSealedClass(open val a: String)
-
-data class SimpleTSC(val b: Int) : TestSealedClass("hey")
-open class MediumTSC(override val a: String, val b: Int) : TestSealedClass(a)
-data class WildTSC(val c: Boolean, val d: String, val e: Int) : MediumTSC(d, e)
-
 data class ExceptionResponse(val message: String)
+
+data class OptionalParams(
+  @KompendiumParam(ParamType.QUERY) val required: String,
+  @KompendiumParam(ParamType.QUERY) val notRequired: String?
+)
